requirements.txt
keras==2.7.0
matplotlib==3.4.3
numpy==1.21.4
opencv_python==4.5.4.58
pandas==1.3.4
Pillow==8.4.0
scikit_learn==1.0.1
seaborn==0.11.2
tensorflow==2.7.0
tensorflow_addons==0.14.0
